Add additional interfaces for requesting schedules
To better align with polling requirements of the server job, the
schedule api has been updated to include a method for fetching one or
more schedules by station id.

Requesting using the station id is much more straightforward, since it
is the key we iterate by on the server. In the previous interface, the
station name was used for indexing and retrieving the station id. In
order to provide the station name, the station id would have to be used
to index the related station document. This created a redundant process.